Zipper: updates in preparation for user list improvements
- The zipper now uses data-clist internally rather than re-implementing
circularity behavior
- A 'Zipper a' is now a 'Zipper a b' and now zips over a list of trees,
each tree having a label of type 'a' and a list of values each of type
'b'. The zipper only ever focuses on 'b' values. This change is to
support having different groupings in the zipper that have different
presentations, such as a 'channel' group and a 'user' group.
- This change also goes ahead and adds a new type for sidebar group
headers, ChannelListGroup. This change also updates the zipper
construction to use that type for the group labels, but this change does
*not* update rendering to use those values yet.