MMNames: remove _cnChans field in lieu of canonical csChannels data
This change removes am MMNames field that was only ever used to build
the channel zipper, and whose value had to be maintained consistently
with other state, leading to an error-prone and redundant design. This
change removes _cnChans in favor of building the sorted channel name
list directly from the csChannel state.