Stop using sortedUserList in attemptCreateDMChannel
This change required some cleanup in how users are resolved by
nickname. This change added

- Types.userByNickname
- Types.userIdForNickname

This change also changed Types.Users.findUserByNickname to take a Users,
not a [UserInfo], to be more symmetric with other functions in the API.

It also changed channelIdByUsername to do what it says: look up the DM
channel ID only by username. Previously it was looking up by nickname if
that setting was enabled, and that behavior is best handled in another
way.