Zipper: add safe and unsafe focus calls
The zipper is expected to always be non-empty when it is used for the
channel list sidebar, in which case we assume that we will always
be able to get the focused channel from the zipper. However, this means
that the zipper cannot be used in cases where it *should* be able to be
empty, such as completion alternatives or other uses we haven't thought
of yet. The internals support an empty zipper but the public API
doesn't, so this change improves that slightly.

One idea would be to just make 'focus' return a 'Maybe b', but that
would mean that the various places we'd like to assume a non-empty
zipper will now have to deal with the 'Nothing' case rather messily.
Instead, this change makes the 'focus' API "safe" by returning a 'Maybe
b', and adds an "unsafe" API, unsafeFocus :: Zipper a b -> b, that can
be used when this assutmpion is relevant at the call site. That way, the
use of this unsafe API is now more apparent, and uses that don't require
a non-empty zipper can use the safe API.