Draw channel list sidebar from zipper only
This change updates the sidebar rendering logic to use the csFocus
zipper as its source rather than using separate data sources. This
change also updates channel selection mode to base its data set on that
same zipper, by filtering the channel list zipper based on matching
the input pattern. This change also simplified channel selection mode
since previously that was maintaining what was essentially zipper logic,
and now the Zipper is used to take care of that.

It's worth mentioning that this greatly simplified the channel list
rendering, but one reason it got simpler was because the viewport
height optimization was removed. We can always reinstate that if it
becomes relevant, but that's going to be less important once the sidebar
stops showing all users (which has not yet happened).