refreshChannelsAndUsers: stop fetching all users, bulk-fetch DM users (#428, #433)
This change stops fetching all users whenever we refresh channels and
users. This means we rely on lazy loading of user data whenever user IDs
are encountered in data from the server. This introduces a regression,
which is that since we now only fetch users corresponding to DM
channels in this function, users mentioned by name ("@...") in messages
will not get highlighted if their user records have not yet been
fetched, and we will not yet have a way to know that we need to fetch
them.

This change also bulk-fetches the user records associated with the
collection of DM channels, which drastically reduces the startup request
count and improves responsiveness.