deleteMessage: move current channel view check to Events
This change makes deleteMessage responsible only for what its name
suggests: deleting a message in a given channel and updating that
channel's state accordingly. Previously it checked to see if that was
the current channel and then updated its view state, but that's a
separate concern that callers of deleteMessage should take care of
instead. The only call site was in Events, so that call site takes care
of that now.