Update the sidebar when conditions are right
This change adds sidebar update calls to various places where it's
possible that changing conditions should change whether a channel
appears in the sidebar. This change also adds a new ChannelInfo field,
_cdSidebarShowOverride, which takes precedence over other metadata when
deciding whether to include a channel in the sidebar. This is important
because we need a client-side notion of whether a channel should be in
the sidebar that is independent of server metadata since there are times
when we need to show the channel but cannot yet update the metadata (a
bootstrapping issue).