Make asynchronous work API return an *optional* event loop injection
Previously, any asynchronous work always had to return an action of type
'MH ()' that would be injected into the main event loop. In many cases,
that action was a no-op, 'return ()'. That led to unnecessary redraws in
the main event loop. This changes the async API so that async work can
return a 'Maybe (MH ())' where 'Nothing' indicates that no event loop
injection is required.