Only fetch status updates for users in the sidebar (fixes #429)
This change does a few important things:
* Changes the handleNewChannel and refreshChannel functions to take a
boolean argument indicating whether the caller would like these
functions to refresh the sidebar. This is because in most cases we want
these functions to update the sidebar, but in some uncommon cases we
want to batch these operations and only update the sidebar once. This
gives the caller control over that behavior.
* Replaces the MVar/TVar that were used to communicate a user ID list to
the status update thread with a TChan. Whenever the sidebar changes, we
send the new sidebar zipper over the channel.
* As a consequence of the above, we now only fetch status information
for users that appear in the DM section of the sidebar, since those are
the only users whose statuses matter for display purposes.
* This change modifies the user status thread to only fetch when EITHER:
  1) It times out (currently at 30 seconds, same as before this change)
  2) It gets a new zipper update AND the new zipper's user IDs differ
     from the list used in the previous fetch
The above behavior ensures that we only fetch when there's something
worth fetching, or when we have been idle long enough.