Add State.Users.withFetchedUser{,Maybe} (relevant to #448)
This change updates addUserToCurrentChannel,
removeUserFromCurrentChannel, and changeChannelByName to use a new
function that deals gracefully with user records missing from the state.
The idea is that we don't always know about all users, so we have to
attempt a local state lookup and then ask the server if the local
lookup fails. The withFetchedUser* functions added in this commit
provide a convenient way to do just that, and this change updates a few
places that were previously relying on user records to use this new
fetching combinator instead.