Update pending channel changes to support user IDs, too
This change simplifies DM channel creation and removes duplicate channel
creation code by making it so that when we create a DM channel, we rely
on the side effects (and the corresponding websocket event) to update us
asynchronously. Before, the createDMChannel function did a lot of work
that was also done elsewhere. Now, it only does what it says: it creates
a DM channel. Then the server notifies us that we got added to it, at
which point the normal channel management logic kicks in. To ensure
that Matterhorn schedules a pending channel change to that channel, this
change augments the PendingChannelChange type with a constructor for
identifying DM channels by User ID. This change then uses that to
schedule a change to the DM channel once it exists, doing this by user
ID instead of channel ID since we don't know its channel ID at the time
we need to schedule the switch.