Manage group channels the same way DM channels are managed
This change undoes some old hacks that removed group channels from the
state when they were hidden. Those hacks were in place because we
previously had no distinction between channels we knew about that
*weren't* in the sidebar, and those that were. Now that we do and now
that we have filtering that we do when building the sidebar, we can now
do group channel filtering the same way we now do for DM channels: check
various conditions, including a hide/show preference, to determine which
group channels appear in the sidebar.