Indicate unread channels in channel list group headings
This change introduces a new theme attribute,
'unreadChannelGroupMarker', that styles an asterisk that appears to the
right of any channel list heading whose channels have unread activity.

This change also refactors applyPreferenceChange so that if a DM or
group channel preference changes to hide, any local channel list
override for that channel is cleared. This makes it so that such
overrides don't keep those open after a preference has explicitly
indicated that they should be hidden.

This change also updates setFocus so that it *always* sets the local
show override on any channel it switches to. This ensures that if a
channel initially appears in the sidebar because it has unread (but very
old) activity, the channel continues to be visible once the channel has
been read.