Track pending completion requests so we can avoid stale async updates
This change tracks the most recent search string used to issue a
(potentially asynchronous) completion request. We need this because the
async requests can take long enough to complete that the editor could
have changed in the mean time, in which case we really don't want to
show the (now stale) completion results since the editor contents have
changed.