setFocusWith: set cdSidebarShowOverride for new focus
This change fixes a bug where a group channel that appeared in the
sidebar due to having unread messages would disappear from the sidebar
as soon as it was switched to. The reason was that the sidebar update
code no longer had a reason to keep it in the sidebar, which is a bug
because when we switch to channels we should always set the sidebar
override so that they stay in the sidebar even if those other conditions
change. However, we were only doing this for channels using setFocus
where we knew the channel ID in advance, rather than with setFocusWith,
where we don't know the ID in advance but instead use a predicate to
search the sidebar for a matching channel. This change makes it so that
no matter what, uses of setFocusWith guarantee that if a channel is in
the sidebar when setFocusWith is called, it is in the sidebar after the
focus change is complete, too.