updateSidebar: call fetchVisibleIfNeeded on newly-focused channel (fixes #458)
This change makes updateSidebar call fetchVisibleIfNeeded on the
newly-focused channel if it differs from the one focused prior to the
zipper rebuild. This is important because channels entered this way may
not be loaded, and this state transition needs to trigger a check for
unloaded channels.

This change also moves updateSidbar to State.Channels.