Add /reconnect to force websocket reconnection (fixes #405)
This change adds a websocket thread ID to ChatResources so we can kill
that thread whenever we reconnect the websocket. This ensures that we
have at most one thread connected at a time.

This is a minimal change to add this feature, but ultimately this
functionality needs to be rewritten to use Async.