Use constructor matching instead of equality for Message type handling.
By using functions to determine the type of message based on intent
and usage (e.g. isReplyable, isGap, isEditable) instead of direct
equality comparisons on the ADT, it is easier to update the ADT
membership and control the intended effects of that update.  This also
has the benefit of capturing the intended functional aspect of message
handling in a contextual function name.

There is still some potential missed analysis of updating the
MessageType ADT due to catch-all conditions in case statements, but
since most of those case statements appear locally here behind these
intent-based functions, all of the locations needing update
considerations are nearby and more easily updated if necessary.