Update selected message when adding new messages to channel history.
If the newly added messages remove a message that was the previous
selection, the selection should be moved or removed, depending on what
was previously selected.  This gives special handling for removed
gaps, allowing for a future commit that will allow fetching to fill
locations held by these gap messages.