Use different events for filling gap and viewing messages.
This allows different keys to be used for these events.

Previously the same key was used for both, which created an
unsatisfactory behavior when the last gap was completely filled and
the key was hit again, viewing a message (different behavior) and from
which the only exit is back to the Main mode at the bottom of the
scroll region.