Add per-channel ephemeral editor settings (just multiline for now)
This change paves the way to fix #445 by adding a new type,
EphemeralEditState, and then putting one of those in each ClientChannel
as well as the ChatState's editor state. The idea is that one of these
values will ultimately contain all per-channel editor state that we want
to swap out/in on channel changes, to reduce the amount of per-channel
maps we have in a few locations for e.g. history input position, etc.