Events.Main: Page Up now selects the most recent message
Previously, Page Up in Main mode entered message selection mode and also
moved the cursor up by a number of messages. This was jarring since the
amount of scrolling that achieved would depend entirely on the height of
the messages involved. This resulted in very jarring scrolling behavior
and sometimes meant that you'd have to move the cursor back down to
messages bypassed by the initial scrolling.

This change makes it so that the initial entry to message selection mode
triggered by this binding does not do any scrolling.