Abort from run.sh if no executables found.
In the original version, the results of the find operation were sorted
by date to find the most recent executable to run.  However, if find
returned nothing, then the SORTED results were the contents of the
current directory, which it then tried to exec one at a time.  This
version aborts if the find doesn't return anything.