Ensure lack of dist directory doesn't abort run.sh script.
The last operation executed as part of the DIRS loop makes the return
value for the loop.  Since this is now a separate evaluation, the -e
setting for the main loop will abort if the last directory does not
exist.  This change always returns success from this search, even if
there are no directories found.