Use new brick API to plumb Vty state through login and team selection
This change uses a new brick API that allows us to avoid resetting the
terminal in between the various brick applications (login and team
selection) that we use on startup before we start the main application.
Previously, in between applications the terminal would get reset,
displaying the user's original terminal state in between login attempts,
for example. Now that no longer occurs and the terminal state is
not reset in between applications, resulting in a smoother experience
during startup.