Coalesce replies in the same thread in normal mode
This change scans the message sequence to analyze the thread
relationships of adjacent messages and use that analysis to affect how
thread replies are rendered as per #490.

This change does not update message select mode to behave the same as
normal channel rendering mode. That change will happen in a separate
commit.