Use post prop "from_webhook" to indicate "BOT" messages (fixes #494)
This change propagates the "from_webhook" property data through
ClientPosts and Messages to determine whether the "[BOT]" indicator is
displayed on message author strings. Previously we looked at the
username override prop, but that is orthogonal to whether the message is
considered to have originated from a bot.

To make this work, this change propagates the bot flag into various
data structures, culminating in checking the flag when a message is
rendered and then emitting the bot suffix at render time. Previously we
appended the bot indicator to the username at message *construction*
time, which meant that later on, when we needed to determine the
username color for that message, the bot suffix text was included in the
username hash input! This change decouples the bot status from the text
of the username and also updates the rendering so that the bot suffix
does not get rendered in the same color as the username.