UserListOverlayState: rename to ListOverlayState, make polymorphic
This change is the first step in an attempt to generalize the user list
overlay machinery enough to make it usable to also handle channel
listings and searches. This change just changes the name, type, and
overlay field names of the overlay state type. It does not change the
functionality at all.