Add tab-completion for standard and custom emoji
This change adds the "emoji.json" file from the upstream project to the
source tree. Matterhorn loads this file on startup (details on this in
the README) and uses the data to build a trie for autocompletion in the
editor. This change also adds the emoji file (and a license) to
releases.