Autocomplete: add context option to complete first match
This change adds a field, autocompleteFirstMatch, to the
AutocompleteContext. If set to True, this field causes the autocomplete
logic to automatically replace the user's input with the first available
match once the matches are available (from the server, say). If set to
False, no such replacement occurs and the autocomplete UI appears with
no initial match selected (the previous behavior).

This change was made to support a more reasonable emoji completion
behavior, where ":<TAB>" previously just showed the autocomplete UI, but
the expected behavior was to also select the first match and complete
the input with it.