Remove duplication of logging output.
When logging to a file target is stopped and then later resumed to
that same target, the start of logging is associated with dumping the
entire contents of the internal log buffer.  This could result in
duplicating previously logged messages when the non-logging period was
relatively short and causes confusion in reading the log file.

This change remembers the last message logged to (the last 50) log
targets and on starting logging to a file, it only logs messages in
the buffer that have not already been logged.