Use shorter term for log messages not context specific (i.e. global).
It is possible to specify a context for a log message (e.g. a
particular channel ID).  The implication of this is that when there is
no context the message is "global".  This simply changes the output
designation to a single character for these messages to economize on
logging line length while preserving the number and order of elements
on the line.