ViewMessage: wrap text blocks while preserving code blocks
Previously the message view window would just turn off text-wrapping
entirely, which was pretty awful for long messages because they'd turn
into single-line messages that needed to be scrolled to be read. We
didn't have wrapping on then, because turning it on would also have
meant limiting the wrapping column and truncating the code blocks - also
a non-starter for the message view window.

This change combines the best of both worlds by *putting off wrapping*
until the Markdown is converted into widgets on a block-by-block basis.
The caller of the message renderer gets to decide whether to wrap
non-code blocks, and if so, at what width. The render does that, but
refrains from truncating code blocks, so the resulting image has text
blocks wrapped but code blocks preserved at their original width. This
allows the message view window to place that resulting image in a
(cached) viewport and then pan around it.