ViewMessage: get the latest version of the message to be displayed (fixes #526)
This change makes it so that the ViewMessage window always tries to get
the latest message from the channel in case it has changed and made the
version in _csViewedMessage stale. That can (and does) happen when the
Message in that state field corresponds to a user message; previously
edits or reactions would not manifest in the View Message window because
it always used the version of the message as it existed at the time the
window was opened.

This change makes it so that we *do* use that message it has no ID,
because that kind of message corresponds to a client messsage or
something that *can't* change over time. But if it has an ID, we always
get the latest version by looking for it in the current channel (where
it is assumed to come from) and use that for rendering.

As part of the change, this patch makes it so that the window rendering
no longer caches the rendering of the tab panes (the message itself, and
the reaction list). This is because invalidating those rendering cache
entries to ensure the latest message is displayed got a lot more
difficult now that we are getting the latest message. And since the
benefit of caching those renderings was marginal, I figured it would be
simplest overall if we just didn't bother caching.