Provide "writeBChan" implementation that shuts the application down on a full queue (relevant to #528)
This change introduces a new function, "writeBChan", that deliberately
clashes with the one provided by Brick. The one added here is in the
Types module, which everything imports, which virtually guarantees that
any module intending to use the Brick version will get a compiler error
about a clash.

The new version raises an exception (and brings the application down) if
the event queue is full. We do this because this is really a
pathological case where we're producing events faster than we can handle
them (such as in a tight loop) or due to a bug.