Draw.Main: preserve date transition lines prior to oldest messages (fixes #534)
This change fixes a long-standing bug where the date line would not
appear above the oldest message in a channel. It turned out that the
date line *was* in the channel message list, because the routine that
inserts the date transition lines is actually correct! The problem was
that the functions to insert thread states both aborted the message
traversal once they encountered a user message with no user message
preceding it in the message history. This meant the date transition line
was getting dropped from the message list just prior to rendering.