Support generation of command tables (fixes #532)
This change adds support for generating plain text and markdown
renderings of the command tables, similar to how we generate keybinding
tables. The `-m` and `-M` options generate the tables for text and
markdown, respectively. Similar to keybindings, the latest table is in
revision control at `docs/commands.md` and is checked by `mkrelease.md`
and is linked from the README.