Show channel display names, not URL names, in sidebar (fixes #522)
This change removes channel sigils from the sidebar and shows channel
display names rather than URL names. This is motivated not only by
a desire to be consistent with the web client but also to avoid the
problem that "/rename" previously had no apparent effect in Matterhorn
due to changing the display name rather than the name we showed.

This change has no impact on how channels are matched in autocompletion,
which is still capabale of matching against anything related to a
channel.