Support return to original channel after M-a sequence (fixes #572)
This change introduces a new behavior: when a user is in channel A and
sees unread activity in channels B and C, the user can press M-a (go to
next unread channel) to go to B, then press it again to go to C, then
press it again to return to A. While the user is in B and C, A will be
marked with a new sigil in the sidebar ("~") to indicate that it is
the "home" channel that the user will return to after reading unread
messages.

While the user is in B or C, any editing keypress or channel-changing
action other than M-a will reset the home channel status and break the
cycle.