Support customization of the attribute and style for the current user (fixes #514)
This change introduces a new theme attribute, 'currentUser', that can be
customized in user theme files to change how the currently logged-in
user's username is rendered. By default, the built-in themes still use
the built-in color hash-bucket approach to color the current user's
username. With this change, the built-in themes also apply a bold font
style to the current user. Any theme customizations will apply on top of
the built-in themes' behavior.