Highlight a channel as unread if a new message is a join or leave
This change fixes the behavior of join/leave messages. Previously a
channel that got a join/leave message did not appear as unread in the
sidebar, despite the client still triggering the activity bell. This led
to the client appearing to have no unread channels, which was wrong.

This change modifies 'adjustUpdated' so that it *always* updates the
channel's update timestamp (rather than ignoring updates for join/leave
messages). That pushes the responsibility to check whether to skip an
update timestamp change to the caller. Now, addMessageToState skips the
call to adjustUpdated if the message was a leave/join *and* if the
user's preferences indicate that they don't want to get notified about
those. This fixes the sidebar behavior; such channels now appear unread
as expected.