Autocomplete: track the type of completion
This change fixes a bug that arises when an active completion search
transitions from one type to another. Previously the logic would
potentially use a cached set of completions from one type (e.g.
channels) as the result for another type (e.g. users) and that would
lead to confusion and a broken UI.

This change tracks the kind of completion that is active; if a cache
lookup for results of a *new* type occurs, the cache lookup always
misses, forcing the new type of completion to trigger a data query (e.g.
server API or local lookup).