Make most editing keys rebindable
This change adds new key events for every binding in the main message
editor's keybinding list and updates the list to use those key events
rather than hard-coding the keybindings themselves. This is important
because it makes it possible to rebind special keys like C-f which could
not previously be customized by users wanting alternatives.

It's important to note that these key events *only* impact the behavior
of the main UI's message editor. They do not have any effect on line
editors used anywhere else in the program. Previously that was perhaps
not such a big deal, but as soon as a user customizes one of these
bindings and finds that it doesn't work as expected in the other
editors, we might have a problem. So some future work will be to
generalize this machinery to work for any editor in the application.