Generalize editor keybindings to work with all editors
This change makes it so that the editing keybindings used with the main
message editor can be used to handle events in *any* (Brick) editor as
long as that editor has a corresponding lens,

  Lens' ChatState (Editor Text Name)

This change re-uses the main editor bindings to work with the editors in
the list overlay window (used for e.g. /join) as well as channel
selection mode (C-g). This makes it possible for customized editor
bindings to apply everywhere in the application that we provide a text
editor.

This change also modified the handleKeyboardEvent function so that it
returns a boolean indicating whether it handled an event by finding a
matching keybinding (True) or whether it failed and just ran the
fallback action it was given (False). We need that boolean value to make
sure we do the right thing to handle the event in some situations.