Refactor key event handling
This patch is quite large but does only a couple of fundamental things.
The goal of this patch is to improve the abstractions we use to express
how key events are handled, both in the abstract as well as in the
presence of custom keybinding configuration from the user (KeyConfig).

One of the motivations for this work is that prior to this patch, there
was no way to statically determine which key events we handled *unless*
we provided a user KeyConfig and built up a list of handlers. This tight
coupling meant that there was a lot of messiness in the data types for
expressing the various situations (handlers that have no default
bindings, handlers that have been explicitly unbound by the user
configuration). It also meant that it was not possible to generate help
text from the configured key handlers while *also* still being able to
determine what the handler collections looked like prior to
configuration. It furthermore prevented another change I wanted to
explore, which is to pre-configure and build up all of the key handlers
and keep them around rather than building them on every single event.
This decoupling change makes that change possible.

For now, though, this change should not change any functionality. The
only user-facing changes should be minor cosmetic changes in how the
help text is rendered in various places.

In terms of details:

* Got rid of the Keybinding type and decomposed it into EventHandler,
KeyEventHandler, and KeyHandler. The various types capture the
situations we need to track and model, where handlers may or may not be
bound to events, and may or may not be associated with specific keys
(Vty events).

* Split up all keybinding lists in the various modules into KeyHandler
lists (where the specific bindings in effect in the configuration, as
well as the defaults, have been populated and mapped to the various
handlers) and handler lists (the abstract event and static key handlers
as they exist prior to configuring them from the user). The former are
used for event routing; the latter are used for help text generation. In
the future the former may be entirely moved into the ChatState to avoid
having to rebuild them on each event dispatch.