Parameterize built-in themes on list of username colors to use
This change decouples the number of username color hash buckets from the
list of colors used to populate those buckets. This way, the color
sequence for a given theme can just be re-used (cycled) to fill in all
hash buckets. This allows us to have different numbers of username
colors by theme, while keeping the number of hash buckets constant
across themes. That makes user theme configurations more stable and it
makes the help documentation easy to write. Lastly and most of all, it
paves the way for themes to use different color lists as their username
colors (such as lists of 256-color palette entries).