ListOverlayState: carry total record count and display it in the title bar (fixes #645)
This change extends the ListOverlayState so that it can carry an
optional total record count (such as total known number of users in a
channel). The change also updates the drawing code to display the count
in the title of the pop-up window. This change also improves the status
message that gets shown to communicate the results shown.

It's worth mentioning that this change only adds a total count to the
"/members" overlay since an API function for that data was already
available. No other overlays yet feature total counts.