Simplify abstraction for opening URLs and paths
This change removes the OpenInBrowser type and factors openURL into
openLinkChoice and openFilePath. This allows us to get away from the
messy abstraction of "things that can be opened with the opener command"
that made the assumption that the URL opener command would be used to
open all things. Future support for permalinks would have violated that
assumption, leading to an awkward implementation.