Draw.RichText: fix code block rendering in message view window (fixes #666)
This change exposes control over when to wrap non-highlighted code
blocks via the MessageData field mdWrapNonhighlightedCodeBlocks. We need
this because in the message view window we never want to line-wrap such
blocks, but everywhere else, we do. This change ensures that the
wrapping is done when desired elsewhere, but not in the message view
window case.

As a side note, this change is the latest in a long series that show how
the Draw.RichText parameter passing is getting out of hand. It is time
for a reader monad to tuck those things away.