RichText: add angle bracket decoration to all hyperlinks
This change is intended to get around issues caused when a hyperlink
label has a style or color that prevents the normal hyperlink color from
being used, thus concealing the fact that it is a link. This makes all
links visible regardless of link label content.

In particular, links labeled "foo" now get rendered as "<foo>". Links
with no labels also get decorated this way.

This change necessitated an enhancement to RichText to allow for AST
nodes that describe text element sequences that cannot be broken up
during line wrapping. This was added to ensure that the "<" added to
the first word of a label is never separated from that word during line
wrapping. For example, a link label "foo bar baz" will get turned into
"<foo bar baz>" and "<foo" and "baz>" will be treated as atomic units
and line wrapping will never break up "<" from "foo".