State.UrlSelect.findURLs: do not consider URLs equivalent if they have the same target but different labels
This change fixes a problem where URLs mentioned in messages did not
appear in the URL list (C-o) because they had the same target as other
links in the channel even though they had different labels. This led to
confusion when it was clear that there were URLs mentioned in the
channel that were omitted from the C-o list. This change makes it so
that de-duplication of URLs also considers the URL label to avoid this
problem.