Draw modules: move main layer composition out of all draw modules into Draw.hs
This change moves all calls to 'drawMain' out of all of the modules that
were making those calls and into Draw.hs. This gives us one place to
manage the decisions we make about which layers of the UI should be
shown in each mode and leaves each drawing module to focus only on the
UI it needs to draw.