Show thread root message in message preview when replying (fixes #670)
This change makes it so that when the user uses C-r or C-s/r to reply to
a post, rather than showing that particular post in the message preview,
the root post of that post's thread is shown. This is intended to be
less surprising than the old behavior since showing the root post in the
preview is consistent with the display of the post in the message list
once it is posted, where the root post is shown above the reply.