Support showing/hiding multi-line channel topics
Previously we displayed all lines in a multi-line channel topic but
replaced newlines with spaces. That resulted in broken layouts for
multi-line topics where layout was important. Now we render the topics
as originally intended with user control over how multi-line topics are
displayed in order to let the user control how the topic affects screen
real estate use.

* Added a new key event, 'toggle-expanded-channel-topics', to toggle
  whether channel topics show only their first line or all lines
* Added a new configuration setting, 'showExpandedChannelTopics', to set
  this behavior on startup (default: True)
* Added a new command, '/toggle-expanded-topics', and a new default
  keybinding, 'F3', to control this behavior at runtime