Make pending channel changes carry MH actions
This change makes it so that when a pending channel change is stored
(PendingChannelChange), it optionally carries an MH value to be
evaluated after the channel in question becomes the selected (focused)
channel. This was necessary to ensure that we can select a message in
message selection mode only after a channel is joined and focused.