Types.RichText: remove nested hyperlink removal
According to the Commonmark specification[1], a hyperlink nested within
the label of another hyperlink takes precedence over the enclosing
hyperlink. As a result, the commonmark package leaves enclosing
hyperlink markup intact rather than parsing it as abstract syntax.
Consequently we do not need the now-defunct logic of doing the opposite
(making the enclosing hyperlink take precedence).

[1] https://spec.commonmark.org/0.29/#links