Add TeamId argument to most Name constructors
This change makes it so that most of our resource names used by Brick
also carry a team ID. This is necessary because we need Brick to track
viewport states (and render cache entries) for each UI element specific
to each team that the user can view. Although the application does not
yet support multiple teams, this makes it possible for us to show an
entirely different UI state (including editor or other viewport scroll
states) on a per-team basis. It's a bit noisy as changes go, and it
leads to a little bit of awkardness when names need to be constructed
since TeamIds are now needed, but it enables the entire UI to be
instanced on a per-team basis.